Fix narrowing and sign conversions

+++ examples/example.c
@@ -184,7 +184,7 @@ main(int argc, char** argv)
         } else if (!strncmp(line, "/historylen", 11)) {
             // The "/historylen" command will change the history len
             int len = atoi(line + 11);
-            comlinHistorySetMaxLen(state, len);
+            comlinHistorySetMaxLen(state, (size_t)len);
         } else if (!strncmp(line, "/mask", 5)) {
             comlinMaskModeEnable(state);
         } else if (!strncmp(line, "/unmask", 7)) {
